backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
cohere==4.9.0
qdrant-client==1.9.1
python-dotenv==1.0.0
google-generativeai==0.4.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3


# requests==2.31.0
# beautifulsoup4==4.12.2
# cohere==4.9.0
# qdrant-client==1.9.1
# python-dotenv==1.0.0
# google-generativeai==0.4.1
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# pytest==7.4.3